+++ demo-projects/meetup/site/pages/about.js
@@ -17,10 +17,12 @@ const { publicRuntimeConfig } = getConfig();
 export default function About() {
   const { meetup } = publicRuntimeConfig;
 
-  const { data = {}, loading, error } = useQuery(GET_ORGANISERS);
+  const { data: { allOrganisers: organiserData = [] } = {}, loading, error } = useQuery(
+    GET_ORGANISERS
+  );
 
-  const hasOrganisers = Boolean(data.allOrganisers && data.allOrganisers.length);
-  const allOrganisers = data.allOrganisers.filter(o => o.user).map(o => o.user);
+  const hasOrganisers = Boolean(organiserData && organiserData.length);
+  const allOrganisers = organiserData.filter(o => o.user).map(o => o.user);
 
   return (
     <>