+++ packages/app-version/index.js
@@ -1,0 +1,3 @@
+const { AppVersionProvider, appVersionMiddleware } = require('./lib/app-version');
+
+module.exports = { AppVersionProvider, appVersionMiddleware };